use u32 for remaining encodable block dimensions

+++ crates/backend-uzu/src/encodable_block/prediction_head.rs
@@ -28,7 +28,7 @@ pub enum PredictionHeadError<B: Backend> {
 }
 
 pub struct PredictionHead<B: Backend> {
-    hidden_dim: usize,
+    hidden_dim: u32,
     activation: ActivationType,
     dense_projection: Box<dyn Linear<B>>,
     activation_kernel: <B::Kernels as Kernels>::ActivationKernel,
@@ -38,8 +38,8 @@ pub struct PredictionHead<B: Backend> {
 
 impl<B: Backend> PredictionHead<B> {
     pub fn new(
-        hidden_dim: usize,
-        num_labels: usize,
+        hidden_dim: u32,
+        num_labels: u32,
         data_type: DataType,
         config: &PredictionHeadConfig,
         parameter_tree: &ParameterTree<B>,
@@ -91,19 +91,13 @@ impl<B: Backend> PredictionHead<B> {
     pub fn encode(
         &self,
         input: Allocation<B>,
-        batch_dim: usize,
+        batch_dim: u32,
         encoder: &mut Encoder<B>,
     ) -> Result<Allocation<B>, B::Error> {
         encoder.push_debug_group("prediction head");
 
         let mut hidden = self.dense_projection.encode(input, batch_dim, encoder)?;
-        self.activation_kernel.encode(
-            None::<&Allocation<B>>,
-            &mut hidden,
-            self.hidden_dim as u32,
-            self.activation,
-            encoder,
-        );
+        self.activation_kernel.encode(None::<&Allocation<B>>, &mut hidden, self.hidden_dim, self.activation, encoder);
         let normalized = self.normalization.encode(&hidden, 0, batch_dim, None, encoder)?;
         let logits = self.readout.encode(normalized, batch_dim, encoder)?;
